refactor: 반복문을 이용하여 생성할 수 있는 자료구조는 함수로 자료 구조 및 데이터를 생성 하는 방식으로 변경
