+++ src/constants/selectOption.ts
@@ -1,5 +1,8 @@
-import { generateNextTwoWeeksDates } from '../helper/dateGenerater.ts'
-import { generateGameTimeObjects } from '../helper/generateGameTimeObjects.ts'
+import {
+  generateGameTimeObjects,
+  generateNextTwoWeeksDates,
+  generateTotalPlayersObjects,
+} from '../helper/generateObjectsArray.ts'
 
 export const OPTIONS = {
   REGION: [
@@ -34,21 +37,7 @@ export const OPTIONS = {
     { NAME: '선택취소', VALUE: '' },
   ],
   GAME_TIME: generateGameTimeObjects(),
-  TOTAL_PLAYERS: [
-    { NAME: '1명', VALUE: '1' },
-    { NAME: '2명', VALUE: '2' },
-    { NAME: '3명', VALUE: '3' },
-    { NAME: '4명', VALUE: '4' },
-    { NAME: '5명', VALUE: '5' },
-    { NAME: '6명', VALUE: '6' },
-    { NAME: '7명', VALUE: '7' },
-    { NAME: '8명', VALUE: '8' },
-    { NAME: '9명', VALUE: '9' },
-    { NAME: '10명', VALUE: '10' },
-    { NAME: '11명', VALUE: '11' },
-    { NAME: '12명', VALUE: '12' },
-    { NAME: '선택취소', VALUE: '' },
-  ],
+  TOTAL_PLAYERS: generateTotalPlayersObjects(),
   ADD_FRIEND: [
     { NAME: '허용', VALUE: 'ALLOW' },
     { NAME: '불가', VALUE: 'DISALLOW' },
